mounting draft only when there is draft
