+++ app/screens/home/channel_list/categories_list/index.tsx
@@ -1,90 +1,23 @@
 // Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
 // See LICENSE.txt for license information.
 
-import React, {useEffect, useMemo} from 'react';
-import {useWindowDimensions} from 'react-native';
-import Animated, {useAnimatedStyle, useSharedValue, withTiming} from 'react-native-reanimated';
+import {withDatabase, withObservables} from '@nozbe/watermelondb/react';
+import {switchMap} from 'rxjs/operators';
 
-import DraftsButtton from '@components/drafts_buttton';
-import ThreadsButton from '@components/threads_button';
-import {TABLET_SIDEBAR_WIDTH, TEAM_SIDEBAR_WIDTH} from '@constants/view';
-import {useTheme} from '@context/theme';
-import {useIsTablet} from '@hooks/device';
-import {makeStyleSheetFromTheme} from '@utils/theme';
+import {observeDraftCount} from '@queries/servers/drafts';
+import {observeCurrentChannelId, observeCurrentTeamId} from '@queries/servers/system';
 
-import Categories from './categories';
-import ChannelListHeader from './header';
-import LoadChannelsError from './load_channels_error';
-import SubHeader from './subheader';
+import CategoriesList from './categories_list';
 
-const getStyleSheet = makeStyleSheetFromTheme((theme: Theme) => ({
-    container: {
-        flex: 1,
-        backgroundColor: theme.sidebarBg,
-        paddingTop: 10,
-    },
-}));
+import type {WithDatabaseArgs} from '@typings/database/database';
 
-type ChannelListProps = {
-    hasChannels: boolean;
-    iconPad?: boolean;
-    isCRTEnabled?: boolean;
-    moreThanOneTeam: boolean;
-};
+const enchanced = withObservables([], ({database}: WithDatabaseArgs) => {
+    const currentTeamId = observeCurrentTeamId(database);
+    const draftsCount = currentTeamId.pipe(switchMap((teamId) => observeDraftCount(database, teamId))); // Observe draft count
+    return {
+        currentChannelId: observeCurrentChannelId(database),
+        draftsCount,
+    };
+});
 
-const getTabletWidth = (moreThanOneTeam: boolean) => {
-    return TABLET_SIDEBAR_WIDTH - (moreThanOneTeam ? TEAM_SIDEBAR_WIDTH : 0);
-};
-
-const CategoriesList = ({hasChannels, iconPad, isCRTEnabled, moreThanOneTeam}: ChannelListProps) => {
-    const theme = useTheme();
-    const styles = getStyleSheet(theme);
-    const {width} = useWindowDimensions();
-    const isTablet = useIsTablet();
-    const tabletWidth = useSharedValue(isTablet ? getTabletWidth(moreThanOneTeam) : 0);
-
-    useEffect(() => {
-        if (isTablet) {
-            tabletWidth.value = getTabletWidth(moreThanOneTeam);
-        }
-    }, [isTablet, moreThanOneTeam]);
-
-    const tabletStyle = useAnimatedStyle(() => {
-        if (!isTablet) {
-            return {
-                maxWidth: width,
-            };
-        }
-
-        return {maxWidth: withTiming(tabletWidth.value, {duration: 350})};
-    }, [isTablet, width]);
-
-    const content = useMemo(() => {
-        if (!hasChannels) {
-            return (<LoadChannelsError/>);
-        }
-
-        return (
-            <>
-                <SubHeader/>
-                {isCRTEnabled &&
-                    <ThreadsButton
-                        isOnHome={true}
-                        shouldHighlighActive={true}
-                    />
-                }
-                <DraftsButtton/>
-                <Categories/>
-            </>
-        );
-    }, [isCRTEnabled]);
-
-    return (
-        <Animated.View style={[styles.container, tabletStyle]}>
-            <ChannelListHeader iconPad={iconPad}/>
-            {content}
-        </Animated.View>
-    );
-};
-
-export default CategoriesList;
+export default withDatabase(enchanced(CategoriesList));
